Alle foutmeldingen toegevoegd!
Alle foutmeldingen van alle endpoints toegevoegd!

+++ StudentenhuisAPI/Controllers/deelnemer_controller.js
@@ -38,15 +38,39 @@ function deleteDeelnemerFromMaaltijd(req, res){
             console.log(error);
             res.json(new apiError("No token or wrong token provided", 401));
         } else {
-            let sql = `DELETE FROM deelnemers WHERE StudentenhuisID = ${req.params.studentenhuisID} AND MaaltijdID = ${req.params.maaltijdID} `;
-            console.log(sql);
-            let query = database.connection.query(sql, (error, results) => {
+
+            let first = JSON.stringify(payload.sub);
+            let payloadId = first.replace(/\D/g,'');
+            console.log(payloadId);
+
+            let sql1 = `SELECT UserID FROM deelnemers WHERE StudentenhuisID = ${req.params.studentenhuisID} AND MaaltijdID = ${req.params.maaltijdID}`;
+            console.log(sql1);
+
+            let query = database.connection.query(sql1, (error, result) => {
+
+                console.log(result);
+
+                let compare = JSON.stringify(result);
+                let compare2 = compare.replace(/\D/g,'');
+                console.log("test");
+                console.log(compare2);
+
                 if(error){
                     console.log(error);
-                } else if(results.affectedRows === 0){
-                    res.status(404).json(new apiError("StudentenhuisID or MaaltijdID is not found", 404));
+                } else if (compare2.includes(payloadId)){
+                    let sql = `DELETE FROM deelnemers WHERE StudentenhuisID = ${req.params.studentenhuisID} AND MaaltijdID = ${req.params.maaltijdID} AND UserID =  ` + payloadId;
+                    console.log(sql);
+                    let query = database.connection.query(sql, (error, results) => {
+                        if(error){
+                            console.log(error);
+                        } else if(results.affectedRows === 0){
+                            res.status(404).json(new apiError("StudentenhuisID or MaaltijdID is not found", 404));
+                        } else {
+                            res.status(200).json("Succesfully deleted!")
+                        }
+                    });
                 } else {
-                    res.status(200).json("Succesfully deleted!")
+                    res.json(new apiError("Conflict, you can't delete this", 409));
                 }
             });
         }
@@ -63,23 +87,51 @@ function createDeelnemer(req, res){
             console.log(error);
             res.json(new apiError("No token or wrong token provided", 401));
         } else {
-            const studentenhuisID = req.params.studentenhuisID;
-            const maaltijdID = req.params.maaltijdID;
-            const UserId = payload.sub.ID;
 
-            let UserID = JSON.stringify(payload.sub);
-            let userId = UserID.replace(/\D/g,'');
-            const test = auth.decodePayload(token);
+            let first = JSON.stringify(payload.sub);
+            let payloadId = first.replace(/\D/g,'');
+            console.log(payloadId);
 
-            let sql = "INSERT INTO deelnemers (`StudentenhuisID`, `MaaltijdID`, `UserID`) VALUES ('"+ studentenhuisID + "', '"+ maaltijdID +"', '"+ userId +"')";
-            console.log(sql);
-            let query = database.connection.query(sql, (error, results) => {
+            let sql1 = `SELECT UserID FROM deelnemers WHERE StudentenhuisID = ${req.params.studentenhuisID} AND MaaltijdID = ${req.params.maaltijdID}`;
+            console.log(sql1);
+
+            let query = database.connection.query(sql1, (error, result) => {
+
+                console.log(result);
+
+                let compare = JSON.stringify(result);
+                let compare2 = compare.replace(/\D/g,'');
+
+                console.log(compare2);
+
+
                 if(error){
                     console.log(error);
-                } else if(results.affectedRows === 0){
-                    res.status(404).json(new apiError("StudentenhuisID or MaaltijdID is not found", 404));
+                    res.json(new apiError("Conflict, you can't be added to this meal", 409));
+                } else if (!compare2.includes(payload)){
+
+                    console.log(compare2.includes(payload));
+                    const studentenhuisID = req.params.studentenhuisID;
+                    const maaltijdID = req.params.maaltijdID;
+                    const UserId = payload.sub.ID;
+
+                    let UserID = JSON.stringify(payload.sub);
+                    let userId = UserID.replace(/\D/g,'');
+                    const test = auth.decodePayload(token);
+
+                    let sql = "INSERT INTO deelnemers (`StudentenhuisID`, `MaaltijdID`, `UserID`) VALUES ('"+ studentenhuisID + "', '"+ maaltijdID +"', '"+ userId +"')";
+                    console.log(sql);
+                    let query = database.connection.query(sql, (error, results) => {
+                        if(error){
+                            res.json(new apiError("Conflict, you can't be added to this meal", 409));
+                        } else if(results.affectedRows === 0){
+                            res.status(404).json(new apiError("StudentenhuisID or MaaltijdID is not found", 404));
+                        } else {
+                            res.status(200).json("You've been added to the meal!")
+                        }
+                    });
                 } else {
-                    res.status(200).json("You've been added to a meal!")
+                    res.json(new apiError("Conflict, you can't be added to this meal", 409));
                 }
             });
         }
