+++ StudentenhuisAPI/Controllers/maaltijd_controller.js
@@ -65,15 +65,33 @@ function deleteMaaltijdById(req, res){
             console.log(error);
             res.json(new apiError("No token or wrong token provided", 401));
         } else {
-            let sql = `DELETE FROM maaltijd WHERE StudentenhuisID = ${req.params.studentenhuisID} AND ID = ${req.params.maaltijdID}`;
-            console.log(sql);
-            let query = database.connection.query(sql, (error, results) => {
+
+            let payloadId = payload.sub;
+            console.log(payloadId);
+
+            let sql1 = `SELECT UserID FROM maaltijd WHERE StudentenhuisID = ${req.params.studentenhuisID} AND ID = ${req.params.maaltijdID}`
+            console.log(sql1);
+
+            let query = database.connection.query(sql1, (error, result) => {
+
+                console.log(result);
+
                 if(error){
-                    console.log(error);
-                } else if(results.affectedRows === 0){
-                    res.status(404).json(new apiError("StudentenhuisID or MaaltijdID is not found", 404));
+                    res.json(new apiError("Conflict, you can't delete this", 409));
+                } else if (payloadId = result){
+                    let sql = `DELETE FROM maaltijd WHERE StudentenhuisID = ${req.params.studentenhuisID} AND ID = ${req.params.maaltijdID}`;
+                    console.log(sql);
+                    let query = database.connection.query(sql, (error, results) => {
+                        if(error){
+                            res.json(new apiError("Conflict, you can't delete this", 409));
+                        } else if(results.affectedRows === 0){
+                            res.status(404).json(new apiError("StudentenhuisID or MaaltijdID is not found", 404));
+                        } else {
+                            res.status(200).json("Succesfully deleted!")
+                        }
+                    });
                 } else {
-                    res.status(200).json("Succesfully deleted!")
+                    res.json(new apiError("Conflict, you can't delete this", 409));
                 }
             });
         }
@@ -154,15 +172,38 @@ function updateMaaltijdById(req, res) {
                 return;
             }
 
-            let sql = `UPDATE maaltijd SET Naam = '` + Naam + `', Beschrijving = '` + Beschrijving + `', Ingredienten = '` + Ingredienten + `', Allergie = '`+ Allergie + `', Prijs = '` + Prijs + `' WHERE StudentenhuisID = ` + studentenhuisID + ` AND ID = ` + maaltijdID + ``; 
-            console.log(sql);
-            let query = database.connection.query(sql, (error, results) => {
+
+            let first = JSON.stringify(payload.sub);
+            let payloadId = first.replace(/\D/g,'');
+            console.log(payloadId);
+
+            let sql1 = `SELECT UserID FROM maaltijd WHERE StudentenhuisID = ${req.params.studentenhuisID} AND ID = ${req.params.maaltijdID}`
+            console.log(sql1);
+            let query = database.connection.query(sql1, (error, result) => {
+
+                console.log(result);
+
+                let compare = JSON.stringify(result);
+                let compare2 = compare.replace(/\D/g,'');
+
+                console.log(compare2);
+
                 if(error){
-                    console.log(error);
-                } else if (results.affectedRows === 0){
-                    res.status(404).json(new apiError("StudentenhuisID or MaaltijdID is not found", 404));
+                    res.json(new apiError("Conflict, you can't update this", 409));
+                } else if (payloadId === compare2){
+                    let sql = `UPDATE maaltijd SET Naam = '` + Naam + `', Beschrijving = '` + Beschrijving + `', Ingredienten = '` + Ingredienten + `', Allergie = '`+ Allergie + `', Prijs = '` + Prijs + `' WHERE StudentenhuisID = ` + studentenhuisID + ` AND ID = ` + maaltijdID + ``; 
+                    console.log(sql);
+                    let query = database.connection.query(sql, (error, results) => {
+                        if(error){
+                            console.log(error);
+                        } else if (results.affectedRows === 0){
+                            res.status(404).json(new apiError("StudentenhuisID or MaaltijdID is not found", 404));
+                        } else {
+                            res.status(200).json("Updated!");
+                        }
+                    });
                 } else {
-                    res.status(200).json("Updated!");
+                    res.json(new apiError("Conflict, you can't update this", 409));
                 }
             });
         }
